[SX126x] Added explicit typecast

+++ src/modules/SX126x/SX126x.cpp
@@ -470,7 +470,7 @@ int16_t SX126x::standby(uint8_t mode, bool wakeup) {
     // send a NOP command - this pulls the NSS low to exit the sleep mode,
     // while preventing interference with possible other SPI transactions
     // see https://github.com/jgromes/RadioLib/discussions/1364
-    (void)this->mod->SPIwriteStream(RADIOLIB_SX126X_CMD_NOP, NULL, 0, false, false);
+    (void)this->mod->SPIwriteStream((uint16_t)RADIOLIB_SX126X_CMD_NOP, NULL, 0, false, false);
   }
 
   uint8_t data[] = { mode };
